Add check so no error when updating

+++ install-binaries.bash
@@ -102,7 +102,9 @@ su - "$SUDO_USER" -c "curl -fsSL https://deno.land/x/install/install.sh | sh"
 
 # install kitty TEST
 su - "$SUDO_USER" -c "curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin"
-ln -s ~/.local/kitty.app/bin/kitty $BIN_INSTALL_DIR/kitty
+if [ ! -f "$BIN_INSTALL_DIR/kitty" ]; then
+    ln -s ~/.local/kitty.app/bin/kitty $BIN_INSTALL_DIR/kitty
+fi
 su - "$SUDO_USER" -c "cp $HOME/.local/kitty.app/share/applications/kitty.desktop $HOME/.local/share/applications/"
 su - "$SUDO_USER" -c "sed -i "s | Icon=kitty | Icon=/home/$SUDO_USER/.local/kitty.app/share/icons/hicolor/256x256/apps/kitty.png | g" $HOME/.local/share/applications/kitty.desktop"
